Use dashes in standalone scripts
We were using underscores and dashes in the script names
inconsistently (somewhere one elsewhere the other).

This moves all the scripts that are mostly meant to be run on their
own (like `get_rhcos_image`) to use dashes (which are easier to type
on most keyboards).

The main numbered scripts are left alone for now.